% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aa_chars.R
\docType{data}
\name{aa_chars}
\alias{aa_chars}
\title{Characteristics of amino acids}
\format{
An object of class \code{data.frame} with 22 rows and 26 columns.
}
\usage{
aa_chars
}
\description{
Characteristics of amino acids
}
\keyword{datasets}
