% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aa_chars_subset.R
\docType{data}
\name{aa_chars_subset}
\alias{aa_chars_subset}
\title{Characteristics of amino acids (subset)}
\format{
An object of class \code{data.frame} with 20 rows and 16 columns.
}
\usage{
aa_chars_subset
}
\description{
Characteristics of amino acids (subset)
}
\keyword{datasets}
