% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blastr.R
\name{blastr}
\alias{blastr}
\title{Execute BLAST from R.}
\usage{
blastr(
  seq,
  program = "blastp",
  database = "refseq_protein",
  time.out = NULL,
  chain.single = TRUE,
  maxhits = 2000
)
}
\arguments{
\item{seq}{Sequence as character string, eg "ACVPTY" or "ATCGGC"}

\item{program}{blastn, blastp, blastx, tblastn, tblastx}

\item{database}{Defaults to refseq_protein  Other options include nt (nucleotide), nr (non-redundant protein), refseq_rna (NCBI Transcript Reference Sequences    DNA), swissprot (Non-redundant UniProtKB/SwissProt sequences    Protein), pdbaa (PDB protein database), pdbnt (PDB nucleotide database)}

\item{time.out}{xxx}

\item{chain.single}{xxxx}

\item{maxhits}{Maximum number of hits to keep.  Passed to BLAST parameter HITLIST_SIZE}
}
\description{
Code is modified from the blast.pdb() function of bio3d to allow different databases to be queried and the maximum number of hits to be changed.
}
\details{
For more information see
General: https://ncbi.github.io/blast-cloud/dev/api.html
Databases: https://ncbi.github.io/blast-cloud/blastdb/available-blastdbs.html for
}
