% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chargaff1951.R
\docType{data}
\name{chargaff1951}
\alias{chargaff1951}
\title{Data from publications by Chargaff 1951}
\format{
A data frame, including:
\describe{
\item{sample_name}{Common name or other descriptor}
\item{spp}{Genus and species or other descriptor}
\item{AvsG}{Ratio of A to G}
\item{TvsC}{Ratio of T to C}
\item{PuvsPy}{Ratio of purines to pyrimidines}
}
}
\usage{
chargaff1951
}
\description{
Data from publications by Chargaff 1951
}
\keyword{datasets}
