% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_alignment.R
\name{clean_alignment}
\alias{clean_alignment}
\title{Print multiple sequence alignment (MSA) to R console}
\usage{
clean_alignment(alignment, minpcnongap, minpcid)
}
\arguments{
\item{alignment}{Alignment object}

\item{minpcnongap}{Minimum percentage of non-gap}

\item{minpcid}{Minimum percentage of identical bases}
}
\description{
From Coglan (2011) A little book of R for bioinformatics.
https://a-little-book-of-r-for-bioinformatics.readthedocs.io/en/latest/#
(Originally called printMsa)
}
