% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dengueseq_fasta.R
\docType{data}
\name{dengueseq_fasta}
\alias{dengueseq_fasta}
\title{dengueseq_fasta}
\format{
An object of class \code{character} of length 1.
}
\usage{
dengueseq_fasta
}
\description{
dengueseq_fasta
}
\keyword{datasets}
