% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dengueseqstartstring.R
\docType{data}
\name{dengueseqstartstring}
\alias{dengueseqstartstring}
\title{Data for computational gene finding exercise}
\format{
An object of class \code{character} of length 1.
}
\usage{
dengueseqstartstring
}
\description{
Data for computational gene finding exercise
}
\keyword{datasets}
