% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diag_show.R
\name{diag_show}
\alias{diag_show}
\title{Print the diagonal of a matrix}
\usage{
diag_show(mat, fill = "", row.names = TRUE)
}
\arguments{
\item{mat}{input matrix}

\item{fill}{what to fill the other part of the matrix (defaults to "")}

\item{row.names}{display row names}
}
\description{
Print the diagonal of a matrix
}
