% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dmshrm_blast_1.R
\docType{data}
\name{dmshrm_blast_1}
\alias{dmshrm_blast_1}
\title{BLAST results obtained from bio3d::blast.pdb().
blast.pdb() only queries the Protein Databank (PDB), which only contains high-question sequences.
In contrast, compbio4all::blastr() is a modification of blast.pdb() which  defaults instead to RefSeq.}
\format{
An object of class \code{blast} of length 3.
}
\usage{
dmshrm_blast_1
}
\description{
BLAST results obtained from bio3d::blast.pdb().
blast.pdb() only queries the Protein Databank (PDB), which only contains high-question sequences.
In contrast, compbio4all::blastr() is a modification of blast.pdb() which  defaults instead to RefSeq.
}
\keyword{datasets}
