% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dmshrm_blast_2.R
\docType{data}
\name{dmshrm_blast_2}
\alias{dmshrm_blast_2}
\title{BLAST results obtained with the blastr() function}
\format{
An object of class \code{blast} of length 3.
}
\usage{
dmshrm_blast_2
}
\description{
blastr() is modified from  bio3d::blast.pdb().  blast.pdb() only quries the Protein Databank (PDB), which only contains high-question sequences.  blastr() defaults instead to RefSeq.
}
\keyword{datasets}
