% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dna_seq_stringset.R
\docType{data}
\name{dna_seq_stringset}
\alias{dna_seq_stringset}
\title{Stringset of four mRNA sequences from dengue fever-related viruses in seqinr format}
\format{
An object of class \code{DNAStringSet} of length 4.
}
\usage{
dna_seq_stringset
}
\description{
From examples used by Coghlan (2011)
}
\keyword{datasets}
