% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doves_PA_3_years.R
\docType{data}
\name{doves_PA_3_years}
\alias{doves_PA_3_years}
\title{PA Mourning doves from 3 years, 2005, 2007 and 2010}
\format{
A data frame
\describe{
\item{Route}{Route number}
\item{Year}{Year of study}
\item{route.index}{}
\item{cuts}{}
\item{State}{}
\item{tot.seen}{}
}
}
\usage{
doves_PA_3_years
}
\description{
...
}
\references{
Seamans et al 2012.  Mourning dove population status, 2012.
}
\keyword{datasets}
