% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ensemblleshmania.R
\docType{data}
\name{ensemblleishmania}
\alias{ensemblleishmania}
\title{Leishmania output from ENSEMBL}
\format{
An object of class \code{Mart} of length 1.
}
\usage{
ensemblleishmania
}
\description{
Leishmania output from ENSEMBL
}
\keyword{datasets}
