% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ensemblpfalciparum.R
\docType{data}
\name{ensemblpfalciparum}
\alias{ensemblpfalciparum}
\title{Plasmodium falciparum-related datasets}
\format{
An object of class \code{Mart} of length 1.
}
\usage{
ensemblpfalciparum
}
\description{
Plasmodium falciparum-related datasets
}
\keyword{datasets}
