% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fasta_cleaner.R
\name{fasta_cleaner}
\alias{fasta_cleaner}
\title{Function to convert a FASTA file stored as an object into a vector}
\usage{
fasta_cleaner(fasta_object, parse = TRUE)
}
\arguments{
\item{fasta_object}{R object with FASTA character data to clean}

\item{parse}{Parse data to character vector?}
}
\description{
Function to convert a FASTA file stored as an object into a vector
}
\examples{

dengueseq_fasta <- rentrez::entrez_fetch(db = "nucleotide", id = "NC_001477", rettype = "fasta")
fasta_cleaner(dengueseq_fasta, parse = TRUE)

}
