% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_ORFs_in_seq.R
\name{find_ORFs_in_seq}
\alias{find_ORFs_in_seq}
\title{Find potential open reading frames (ORFs) in a DNA sequence.}
\usage{
find_ORFs_in_seq(sq)
}
\arguments{
\item{sq}{Input sequence}
}
\description{
By Coghlan (2011) A little book of R for bioinformatics.  https://a-little-book-of-r-for-bioinformatics.readthedocs.io/en/latest/.
Originally called findORFsinSeq()
}
\examples{

# From Coglan 2011
## Toy example
s1 <- "aaaatgcagtaacccatgccc"
find_ORFs_in_seq(s1)

## Dengue virus example
data(dengueseqstartstring)
find_ORFs_in_seq(dengueseqstartstring)

}
