% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_start_stop_codons.R
\name{find_start_stop_codons}
\alias{find_start_stop_codons}
\title{Function to identify potential start and stop codons}
\usage{
find_start_stop_codons(sq)
}
\arguments{
\item{sq}{DNA sequence}
}
\description{
By Coghlan (2011) A little book of R for bioinformatics.
https://a-little-book-of-r-for-bioinformatics.readthedocs.io/en/latest/src/chapter7.html.
The functions was originally named findPotentialStartsAndStops().
}
\details{
Searches a DNA sequence using Biostrings::matchPattern(codon, sequence) to look for start and stop codons.
Currently only uses the standard genetic code.
}
\examples{
# Example sequence from Coghlan 2011
s1 <- "aaaatgcagtaacccatgccc"

# Returns a list with 2 elements
## 1) locations of codons
## 2) The codon
find_start_stop_codons(s1)

}
