% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findPotentialStartsAndStops.R
\name{findPotentialStartsAndStops}
\alias{findPotentialStartsAndStops}
\title{Function to identify potential start and stop codons}
\usage{
findPotentialStartsAndStops(sequence)
}
\arguments{
\item{sequence}{DNA sequence}
}
\description{
Function to identify potential start and stop codons
}
