% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_hmm_seq.R
\name{generate_hmm_seq}
\alias{generate_hmm_seq}
\title{Function to generate a DNA sequence, given a HMM and the length of the sequence to be generated.}
\usage{
generate_hmm_seq(transitionmatrix, emissionmatrix, initialprobs, seqlength)
}
\arguments{
\item{transitionmatrix}{transitionmatrix}

\item{emissionmatrix}{emissionmatrix}

\item{initialprobs}{initialprobs}

\item{seqlength}{seqlength}
}
\description{
By Coghlan (2011)
}
