% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_markov_seq.R
\name{generate_markov_seq}
\alias{generate_markov_seq}
\title{Generate a DNA sequence using a Markov Chain.}
\usage{
generate_markov_seq(transitionmatrix, initialprobs, seqlength)
}
\arguments{
\item{transitionmatrix}{transitionmatrix}

\item{initialprobs}{initialprobs}

\item{seqlength}{Length of sequence to generate.}
}
\description{
By Coghlan (2011) A little book of R for bioinformatics.  https://a-little-book-of-r-for-bioinformatics.readthedocs.io/en/latest/.
}
