% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generateSeqsWithMultinomialModel.R
\name{generateSeqsWithMultinomialModel}
\alias{generateSeqsWithMultinomialModel}
\title{generateSeqsWithMultinomialModel}
\usage{
generateSeqsWithMultinomialModel(inputsequence, X)
}
\arguments{
\item{inputsequence}{inputsequence}

\item{X}{X}
}
\description{
generateSeqsWithMultinomialModel
}
