% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genes_dpc.R
\docType{data}
\name{genes_dpc}
\alias{genes_dpc}
\title{Dataset helpfile header . . .}
\format{
A data frame with 51030 rows and 7 column(s)
\describe{
\item{id}{Describe column id here . . .}
\item{chromosome}{Describe column chromosome here . . .}
\item{length}{Describe column length here . . .}
\item{GCpercent}{Describe column GCpercent here . . .}
\item{strand}{Describe column strand here . . .}
\item{start}{Describe column start here . . .}
\item{end}{Describe column end here . . .}
}
}
\source{
\url{http://www.whereyougotthedata.com/}
}
\usage{
genes_dpc
}
\description{
Short description of data . . .
}
\keyword{datasets}
