% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/human_gene_lengths.R
\docType{data}
\name{human_gene_lengths}
\alias{human_gene_lengths}
\title{Lengths of human genes}
\format{
A data frame with 20290 rows and 1 column(s)
\describe{
\item{gene_length}{Describe column gene_lengthhere . . .}
}
}
\source{
\url{http://whitlockschluter.zoology.ubc.ca/wp-content/data/chapter04/chap04e1HumanGeneLengths.csv"}
}
\usage{
human_gene_lengths
}
\description{
Length of the coding portions (exons) of humans genes.
Originally from Whitlock and Schluter 2nd edition.
}
\keyword{datasets}
