% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/leishmaniagenes.R
\docType{data}
\name{leishmaniagenes}
\alias{leishmaniagenes}
\title{Leishmaniabgenes output from ENSEMBL}
\format{
An object of class \code{data.frame} with 10030 rows and 1 columns.
}
\usage{
leishmaniagenes
}
\description{
Leishmaniabgenes output from ENSEMBL
}
\keyword{datasets}
