% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/leishmaniagenes2.R
\docType{data}
\name{leishmaniagenes2}
\alias{leishmaniagenes2}
\title{Subset of Leishmania genes output from ENSEMBL}
\format{
An object of class \code{data.frame} with 10030 rows and 2 columns.
}
\usage{
leishmaniagenes2
}
\description{
Subset of Leishmania genes output from ENSEMBL
}
\keyword{datasets}
