% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Lycettetal1999.R
\docType{data}
\name{Lycettetal1999}
\alias{Lycettetal1999}
\title{Dataset helpfile header . . .}
\format{
A data frame with 65 rows and 15 column(s)
\describe{
\item{behvr.grp}{Describe column behvr.grp here . . .}
\item{grp.i}{Describe column grp.i here . . .}
\item{behvr.i}{Describe column behvr.i here . . .}
\item{behvr.type}{Describe column behvr.type here . . .}
\item{behvr}{Describe column behvr here . . .}
\item{Bs}{Describe column Bs here . . .}
\item{behvr.type.short}{Describe column behvr.type.short here . . .}
\item{behavr.short}{Describe column behavr.short here . . .}
\item{Ta}{Describe column Ta here . . .}
\item{Go}{Describe column Go here . . .}
\item{Ma}{Describe column Ma here . . .}
\item{Mk}{Describe column Mk here . . .}
\item{Kib}{Describe column Kib here . . .}
\item{Bd}{Describe column Bd here . . .}
\item{comments}{Describe column comments here . . .}
}
}
\source{
\url{http://www.whereyougotthedata.com/}
}
\usage{
Lycettetal1999
}
\description{
Short description of data . . .
}
\keyword{datasets}
