% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_dataset_helpfile.R
\name{make_dateset_helpfile}
\alias{make_dateset_helpfile}
\title{Create template for dataset documentation}
\usage{
make_dateset_helpfile(dataset, dataset_name = "temp")
}
\arguments{
\item{dataset}{Dataframe object.}

\item{dataset_name}{Quoted name of the dataset object}
}
\description{
Given an dataframe, creates a template for a R package documentation file.
}
\details{
Defaults are set to generic template.

NOTE: Does not check if a file of the same name currently exists!
}
