% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_seqs_multinom_mod.R
\name{make_seqs_multinom_mod}
\alias{make_seqs_multinom_mod}
\title{make_seqs_multinom_mod}
\usage{
make_seqs_multinom_mod(inputsequence, X)
}
\arguments{
\item{inputsequence}{inputsequence}

\item{X}{X}
}
\description{
By Coghlan (2011)
}
