% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_viterbi_mat.R
\name{make_viterbi_mat}
\alias{make_viterbi_mat}
\title{Function to generate a DNA sequence, given a HMM and the length of the sequence to be generated.}
\usage{
make_viterbi_mat(sequence, transitionmatrix, emissionmatrix)
}
\arguments{
\item{sequence}{sequence}

\item{transitionmatrix}{transitionmatrix}

\item{emissionmatrix}{emissionmatrix}
}
\description{
By Coghlan (2011)
}
