% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pams.R
\docType{data}
\name{pams}
\alias{pams}
\title{Point-accepted mutation (PAM) data matrix from Dayoff et al. 1978}
\format{
An object of class \code{data.frame} with 20 rows and 21 columns.
}
\usage{
pams
}
\description{
Figure 3.8 page 81 in Pevsner. Figure 80 Dayhoff et al 1978 (?).
Values are the number of observed mutations out of 1572 amino acid substitutions.  Multiplied by 10 as specified in original pubs
}
\details{
As noted by Dayoff et al (1978)  "frational exchanges result when ancestral sequences are ambiguous." (pg 346).
}
\keyword{datasets}
