% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pfalciparumgenes.R
\docType{data}
\name{pfalciparumgenes}
\alias{pfalciparumgenes}
\title{Plasmodium falciparum-related datasets}
\format{
An object of class \code{data.frame} with 5767 rows and 2 columns.
}
\usage{
pfalciparumgenes
}
\description{
Plasmodium falciparum-related datasets
}
\keyword{datasets}
