% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_ORFs_in_seq.R
\name{plot_ORFs_in_seq}
\alias{plot_ORFs_in_seq}
\title{Find and plot potential ORFs}
\usage{
plot_ORFs_in_seq(sq)
}
\arguments{
\item{sq}{DNA sequence}
}
\description{
By Coghlan (2011) A little book of R for bioinformatics.  https://a-little-book-of-r-for-bioinformatics.readthedocs.io/en/latest/.
This function calls find_start_stop_ORFs() thenm builds a plot of the output.
Predictions are based on the standard genetic code.
}
\examples{
# From Coglan 2011
## Toy example
s1 <- "aaaatgcagtaacccatgccc"
plot_ORFs_in_seq(s1)

## Dengue virus example
data(dengueseqstartstring)
plot_ORFs_in_seq(dengueseqstartstring)


}
