% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_sliding_window.R
\name{plot_sliding_window}
\alias{plot_sliding_window}
\title{plot_sliding_window}
\usage{
plot_sliding_window(windowsize, inputseq)
}
\arguments{
\item{windowsize}{Size of window}

\item{inputseq}{Sequence to analyze}
}
\description{
By Coghlan (2011)
}
