% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_potential_starts_and_stops.R
\name{plot_start_stop_codons}
\alias{plot_start_stop_codons}
\title{plot_potential_starts_and_stops}
\usage{
plot_start_stop_codons(sq)
}
\arguments{
\item{sq}{sequence to analyze and plot}
}
\description{
From Coghlan (2011) A little book of R for bioinformatics.
https://a-little-book-of-r-for-bioinformatics.readthedocs.io/en/latest/src/chapter7.html
Function original named plotPotentialStartsAndStops()
}
\examples{
# Toy Example sequence from Coghlan 2011
s1 <- "aaaatgcagtaacccatgccc"
plot_start_stop_codons(s1)

# Dengue virus example from Coghlan 2011
data(dengueseqstartstring)
plot_start_stop_codons(dengueseqstartstring)

}
