% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_msa.R
\name{print_msa}
\alias{print_msa}
\title{Print multiple sequence alignment (MSA) to R console}
\usage{
print_msa(alignment, chunksize = 60)
}
\arguments{
\item{alignment}{Alignment object}

\item{chunksize}{width of alignment to output}
}
\description{
From Coglan (2011) A little book of R for bioinformatics.
https://a-little-book-of-r-for-bioinformatics.readthedocs.io/en/latest/#
(Originally called printMsa)
}
