% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_pairwise_alignment.R
\name{print_pairwise_alignment}
\alias{print_pairwise_alignment}
\title{printPairwiseAlignment}
\usage{
print_pairwise_alignment(alignment, chunksize = 60, returnlist = FALSE)
}
\arguments{
\item{alignment}{Alignment to print in console}

\item{chunksize}{chunksize}

\item{returnlist}{returnlist}
}
\description{
From Coghlan (2011) A little book of R for bioinformatics.
https://a-little-book-of-r-for-bioinformatics.readthedocs.io/en/latest/src/chapter4.html
}
\details{
This functions uses Biostrings::pattern() and IRanges::subject() from Bioconductor.
}
