% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/printPairwiseAlignment.R
\name{printPairwiseAlignment}
\alias{printPairwiseAlignment}
\title{printPairwiseAlignment}
\usage{
printPairwiseAlignment(alignment, chunksize = 60, returnlist = FALSE)
}
\arguments{
\item{alignment}{alignment}

\item{chunksize}{chunksize}

\item{returnlist}{returnlist}
}
\description{
printPairwiseAlignment
}
