% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/robinson_aafreq.R
\docType{data}
\name{robinson_aafreq}
\alias{robinson_aafreq}
\title{Dataset helpfile header . . .}
\format{
A data frame with 20 rows and 5 column(s)
\describe{
\item{X}{Describe column X here . . .}
\item{aa3}{Describe column aa3 here . . .}
\item{aa1}{Describe column aa1 here . . .}
\item{aa.count}{Describe column aa.count here . . .}
\item{aa.freq}{Describe column aa.freq here . . .}
}
}
\source{
\url{http://www.whereyougotthedata.com/}
}
\usage{
robinson_aafreq
}
\description{
Short description of data . . .
}
\keyword{datasets}
