% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rooted_NJ_tree.R
\name{rooted_NJ_tree}
\alias{rooted_NJ_tree}
\title{Wrapper function for creating rooted neighbor-joining trees}
\usage{
rooted_NJ_tree(alignment, og, type)
}
\arguments{
\item{alignment}{Multiple sequence alignment object}

\item{og}{Outgroup, as quoted string}

\item{type}{Type of sequencE: "DNA" or "protein"}
}
\description{
By Avril Coghlan (2011).
}
\details{
Uses seqinr::dist.alignment() and ape::nj(), ape::boot.phylo().
}
