% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/score_alignment.R
\name{score_alignment}
\alias{score_alignment}
\title{Score a simple alignment}
\usage{
score_alignment(
  seq.df,
  seq1,
  seq2,
  sub.mat,
  gap.penalty = -10,
  gap.extend = -4
)
}
\arguments{
\item{seq.df}{Two aligned sequences in a dataframe}

\item{seq1}{name of column with first sequences}

\item{seq2}{name of column with second sequence}

\item{sub.mat}{substitution matrix to use.  Defaults to BLOSUM62}

\item{gap.penalty}{penalty for gaps in the sequence}

\item{gap.extend}{penalty for gap length, starting with the first gap.}
}
\description{
Score a simple alignment
}
