% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dengue_seq_1_2_3_4.R
\docType{data}
\name{seq_1_2_3_4}
\alias{seq_1_2_3_4}
\title{seq_1_2_3_4}
\format{
An object of class \code{character} of length 4.
}
\usage{
seq_1_2_3_4
}
\description{
seq_1_2_3_4
}
\keyword{datasets}
