% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dengue_seq_1_2_3_4_stringset.R
\docType{data}
\name{seq_1_2_3_4_stringset}
\alias{seq_1_2_3_4_stringset}
\title{seq_1_2_3_4_stringset}
\format{
An object of class \code{AAStringSet} of length 4.
}
\usage{
seq_1_2_3_4_stringset
}
\description{
seq_1_2_3_4_stringset
}
\keyword{datasets}
