% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shroom_table.R
\docType{data}
\name{shroom_table}
\alias{shroom_table}
\title{shroom_table}
\format{
An object of class \code{data.frame} with 14 rows and 4 columns.
}
\usage{
shroom_table
}
\description{
shroom_table
}
\keyword{datasets}
