% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tri_print.R
\name{tri_print}
\alias{tri_print}
\title{Print the lower or upper triangular part of a matrix  flextable}
\usage{
tri_print(
  mat,
  triangle = "lower",
  fill = "",
  as.image = TRUE,
  row.names = TRUE
)
}
\arguments{
\item{mat}{input matrix}

\item{triangle}{specify upper or lower in quotes}

\item{fill}{what to fill the other part of the matrix (defaults to "")}

\item{as.image}{print it as a image (flextable)}

\item{row.names}{display row names}
}
\description{
Print the lower or upper triangular part of a matrix  flextable
}
