% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tutor_show.R
\name{tutor_show}
\alias{tutor_show}
\title{Show the tutorials currently available for the dayoff package}
\usage{
tutor_show()
}
\description{
Show the tutorials currently available for the dayoff package
}
\examples{
tutor_show()


}
