% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_dayoff.R
\name{update_dayoff}
\alias{update_dayoff}
\title{Function to update dayoff package}
\usage{
update_dayoff(upgrade = "never", type = "binary")
}
\arguments{
\item{upgrade}{Arugement to pass to devtools::install_github: should dependencies be upgrades; defaults to "never}

\item{type}{Arguement to pass to devtools::install_github: should binary packages be downloaded even if more recent source packages are available.  Defaults to "binary."}
}
\description{
Function to update dayoff package
}
