% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/virus_mrna_aln.R
\docType{data}
\name{virus_mrna_aln}
\alias{virus_mrna_aln}
\title{Alignment of four mRNA sequences from dengue fever-related viruses}
\format{
An object of class \code{MsaDNAMultipleAlignment} with 4 rows and 1097 columns.
}
\usage{
virus_mrna_aln
}
\description{
From examples used by Coghlan (2011)
}
\keyword{datasets}
