% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/virus_mrna_aln_bin.R
\docType{data}
\name{virus_mrna_aln_bin}
\alias{virus_mrna_aln_bin}
\title{Alignment of four proteins from dengue fever-related viruses in bin format}
\format{
An object of class \code{DNAbin} with 4 rows and 1097 columns.
}
\usage{
virus_mrna_aln_bin
}
\description{
From examples used by Coghlan (2011)
}
\keyword{datasets}
