% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/virus_mrna_dist.R
\docType{data}
\name{virus_mrna_dist}
\alias{virus_mrna_dist}
\title{Distance matrix of Dengue virus mRNA sequences}
\format{
An object of class \code{dist} of length 6.
}
\usage{
virus_mrna_dist
}
\description{
From examples used by Coghlan (2011)
}
\keyword{datasets}
