% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/virus_snc.R
\docType{data}
\name{virus_snc}
\alias{virus_snc}
\title{Dataset helpfile header . . .}
\format{
A data frame with 39 rows and 15 column(s)
\describe{
\item{nucleic_acid}{Describe column nucleic_acid here . . .}
\item{strands}{Describe column strands here . . .}
\item{reverse_transcribed}{Describe column reverse_transcribed here . . .}
\item{group}{Describe column group here . . .}
\item{group_long}{Describe column group_long here . . .}
\item{virus}{Describe column virus here . . .}
\item{name_short}{Describe column name_short here . . .}
\item{genome_size_kb}{Describe column genome_size_kb here . . .}
\item{mutation_rate}{Describe column mutation_rate here . . .}
\item{fraction_indels}{Describe column fraction_indels here . . .}
\item{mutational_target_size}{Describe column mutational_target_size here . . .}
\item{cell_infection_cycles}{Describe column cell_infection_cycles here . . .}
\item{selection_correction_factor}{Describe column selection_correction_factor here . . .}
\item{estimate_relability}{Describe column estimate_relability here . . .}
\item{source}{Describe column source here . . .}
}
}
\source{
\url{http://www.whereyougotthedata.com/}
}
\usage{
virus_snc
}
\description{
Short description of data . . .
}
\keyword{datasets}
