% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/virus_snr.R
\docType{data}
\name{virus_snr}
\alias{virus_snr}
\title{Dataset helpfile header . . .}
\format{
A data frame with 12 rows and 10 column(s)
\describe{
\item{nucleic_acid}{Describe column nucleic_acid here . . .}
\item{strands}{Describe column strands here . . .}
\item{group}{Describe column group here . . .}
\item{group_long}{Describe column group_long here . . .}
\item{virus}{Describe column virus here . . .}
\item{name_short}{Describe column name_short here . . .}
\item{genome_size_kb}{Describe column genome_size_kb here . . .}
\item{mutation_rate}{Describe column mutation_rate here . . .}
\item{mutational_target_size}{Describe column mutational_target_size here . . .}
\item{source}{Describe column source here . . .}
}
}
\source{
\url{http://www.whereyougotthedata.com/}
}
\usage{
virus_snr
}
\description{
Short description of data . . .
}
\keyword{datasets}
