% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/virusaln.R
\docType{data}
\name{virusaln}
\alias{virusaln}
\title{Alignment of four proteins from dengue fever-related viruses}
\format{
An object of class \code{AAMultipleAlignment} with 4 rows and 305 columns.
}
\usage{
virusaln
}
\description{
From examples used by Coglan (201x)
}
\keyword{datasets}
