% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/virusaln_seqinr.R
\docType{data}
\name{virusaln_seqinr}
\alias{virusaln_seqinr}
\title{Alignment of four proteins from dengue fever-related viruses in seqinr format}
\format{
An object of class \code{alignment} of length 4.
}
\usage{
virusaln_seqinr
}
\description{
Alignment of four proteins from dengue fever-related viruses in seqinr format
}
\keyword{datasets}
