% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/virusaln_seqinr_clean.R
\docType{data}
\name{virusaln_seqinr_clean}
\alias{virusaln_seqinr_clean}
\title{Alignment of four proteins from dengue fever-related viruses with poorly aligned portions removed}
\format{
An object of class \code{alignment} of length 4.
}
\usage{
virusaln_seqinr_clean
}
\description{
From examples used by Coglan (201x)
}
\keyword{datasets}
