% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viterbi.R
\name{viterbi}
\alias{viterbi}
\title{By Coghlan (2011), adapted from "Applied Statistics for Bioinformatics using R" by Wim P. Krijnen, page 209}
\usage{
viterbi(sequence, transitionmatrix, emissionmatrix)
}
\arguments{
\item{sequence}{sequence}

\item{transitionmatrix}{transitionmatrix}

\item{emissionmatrix}{emissionmatrix}
}
\description{
By Coghlan (2011), adapted from "Applied Statistics for Bioinformatics using R" by Wim P. Krijnen, page 209
}
