% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wine.R
\docType{data}
\name{wine}
\alias{wine}
\title{Wine recognition data}
\format{
A data frame ....
\describe{
\item{V1}{Cultivar, 1 to 3}
\item{V2}{Alcohol}
\item{V3}{Malic acid}
\item{V4}{Ash}
\item{V5}{Alcalinity of ash}
\item{V6}{Magnesium}
\item{V7}{Total phenols}
\item{V8}{Flavanoids}
\item{V9}{Nonflavanoid phenols}
\item{V10}{Proanthocyanins}
\item{V11}{Color intensity}
\item{V12}{Hue}
\item{V13}{OD280/OD315 of diluted wines}
\item{V14}{Proline}
}
}
\usage{
wine
}
\description{
Downloaded from
http://archive.ics.uci.edu/ml/machine-learning-databases/wine/wine.data
}
\details{
Original source:
Forina, M. et al, PARVUS - An Extendible Package for Data
Exploration, Classification and Correlation. Institute of Pharmaceutical
and Food Analysis and Technologies, Via Brigata Salerno,  16147 Genoa, Italy.

These data are the results of a chemical analysis of wines grown in the same region in Italy but derived from three different cultivars. The analysis determined the quantities of 13 constituents found in each of the three types of wines.
}
\keyword{datasets}
