% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compbio4all-package.R
\docType{package}
\name{compbio4all-package}
\alias{compbio4all}
\alias{compbio4all-package}
\title{compbio4all: Datasets and helpful functions for learning bioinformatics and computational biology}
\description{
What the package does (one paragraph).
}
\author{
\strong{Maintainer}: Nathan Brouwer \email{brouwern@gmail.com} (\href{https://orcid.org/YOUR-ORCID-ID}{ORCID})

}
\keyword{internal}
