% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unrooted_NJ_tree.R
\name{unrooted_NJ_tree}
\alias{unrooted_NJ_tree}
\title{Wrapper function for creating unrooted neighbor-joining trees}
\usage{
unrooted_NJ_tree(alignment, type)
}
\arguments{
\item{alignment}{Multiple sequence alignment}

\item{type}{"DNA" or "protein"}
}
\description{
By Avril Coghlan (2011).
}
\details{
Uses seqinr::dist.alignment() and ape::nj(), ape::boot.phylo().
}
